28925: Modify find_python_sources to support modularization (squashed)

+++ src/sage_setup/clean.py
@@ -15,7 +15,7 @@ Clean the Install Dir
 import os
 import importlib.util
 
-from sage_setup.find import installed_files_by_module, get_extensions
+from sage_setup.find import installed_files_by_module, get_extensions, read_distribution
 
 
 
@@ -87,8 +87,7 @@ def _find_stale_files(site_packages, python_packages, python_modules, ext_module
         ....:     ['sage', 'sage_setup'], cythonized_dir, []).items())
         sage: from sage_setup.clean import _find_stale_files
 
-    TODO: move ``module_list.py`` into ``sage_setup`` and also check
-    extension modules::
+    TODO: Also check extension modules::
 
         sage: stale_iter = _find_stale_files(SAGE_LIB, python_packages, python_modules, [], extra_files)
         sage: from sage.misc.sageinspect import loadable_module_extension
@@ -141,7 +140,8 @@ def _find_stale_files(site_packages, python_packages, python_modules, ext_module
                 yield f
 
 
-def clean_install_dir(site_packages, python_packages, python_modules, ext_modules, data_files, nobase_data_files):
+def clean_install_dir(site_packages, python_packages, python_modules, ext_modules, data_files, nobase_data_files, *,
+                      distributions=None):
     """
     Delete all modules that are **not** being installed
 
@@ -178,5 +178,6 @@ def clean_install_dir(site_packages, python_packages, python_modules, ext_module
         site_packages, python_packages, python_modules, ext_modules, data_files, nobase_data_files)
     for f in stale_file_iter:
         f = os.path.join(site_packages, f)
-        print('Cleaning up stale file: {0}'.format(f))
-        os.unlink(f)
+        if distributions is None or read_distribution(f) in distributions:
+            print('Cleaning up stale file: {0}'.format(f))
+            os.unlink(f)
